# Make ttnn.divide use SFPU to handle division by zero cases