… tests/ttnn/unit_tests/operations/eltwise/test_binary_fp32.py
… tests/ttnn/unit_tests/operations/eltwise/test_binaryng_fp32.py
… tests/ttnn/unit_tests/operations/eltwise/test_divide.py
… ttnn/cpp/ttnn/operations/eltwise/binary_ng/device/binary_ng_device_operation.hpp
… ttnn/cpp/ttnn/operations/eltwise/quantization/quantization.cpp