### Ticket
#25629

### Problem description
The float32 version of `ttnn.divide` uses the divide LLK (SFPU) and handles division by zero cases 
The bfloat16 version uses the `recip + mul` LLKs (SFPU + FPU) but does not handle division by zero.

In the second case, there can be PCC-related issues if one input is 0, because its output value will be ~10**38 rather than infinity. 

### What's changed

This PR makes the bfloat16 version of `divide` use `divide` LLK (like the float32 version) rather than `recip + mul` and make it handle division by 0 cases. 

This allows `divide` to pass tests from #25629 . 

### Performance evaluation

The proposed solution does have a performance cost.
In this case, we get a performance hit: slower by a factor of 1.3-1.5 in the worst case. 

#### Wormhole:
  | Width | Height | Size | Time [ns] | Throughput [GElem.s] | Speedup vs. proposed
-- | -- | -- | -- | -- | -- | --
SFPU divide (fix) | 2048 | 2048 | 4194304 | 237.152 | 17686 |  N/|A
| | 1024 | 1024 | 1048576 | 62.002 | 16912 |  
| | 512 | 512 | 262144 | 17.576 | 14915 |  
recip + mul (SFPU + FPU)  | 2048 | 2048 | 4194304 | 165.792 | 25299 | 1.43
| (current) | 1024 | 1024 | 1048576 | 44.336 | 23651 | 1.40
| | 512 | 512 | 262144 | 13.16 | 19920 | 1.34


#### Blackhole:
  | Width | Height | Size | Time [ns] | Throughput [GElem.s] | Speedup vs. proposed
-- | -- | -- | -- | -- | -- | --
SFPU divide (fix) | 2048 | 2048 | 4194304 | 104.992 | 39949 |  
| | 1024 | 1024 | 1048576 | 26.614 | 39399 |  
| | 512 | 512 | 262144 | 7.24 | 36208 |  
recip + mul (SFPU + FPU) | 2048 | 2048 | 4194304 | 67.274 | 62347 | 1.56
| (current) | 1024 | 1024 | 1048576 | 17.522 | 59843 | 1.52
| | 512 | 512 | 262144 | 5.258 | 49856 | 1.38

These performance have been obtained by looping 500 times on the compute kernel code between circular buffers reserve_back/pop_front/etc... 

These are essentially _worst-case_ slowdowns. In practice, the difference in execution time should be lower due to other overheads (e.g. core-to-core synchronization). We are also comparing an algorithm that handles infinity vs. one that does not.


### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI with demo tests passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable) See [recommended dev flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy quick](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-quick-trigger.yaml) CI passes (if applicable)
- [ ] [TG demo tests, for Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-demo-tests.yaml) CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml) CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] New/Existing tests provide coverage for changes